pyproject.toml
[tool.poetry]
name = "erd-ga"
version = "0.1.0"
description = ""
authors = ["n.desaunais"]

[tool.poetry.dependencies]
python = "^3.10.9"
loguru = "^0.5.3"
click = "^8.1.7"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.7"
mermaid-py = "^0.5.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.2.2"
testcontainers-postgres = "^0.0.1rc1"
requests = "^2.32.3"
flake8 = "^7.0.0"
isort = "^5.13.2"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
